server/pyproject.toml
[project]
name = "hermes-hiring"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]",
    "pydantic>=2.0",
    "pydantic-settings",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg",
    "pgvector",
    "httpx",
    "redis[hiredis]",
    "boto3",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-httpx",
    "prometheus-client",
    "python-multipart",
    "sentence-transformers",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio",
    "pytest-cov",
    "mypy",
    "ruff",
    "pre-commit",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
